requirements.txt
cachetools==2.0.1
click==6.7
flake8==3.5.0
Flask==0.12.2
google-api-python-client==1.6.5
google-auth==1.4.1
google-auth-httplib2==0.0.3
gunicorn==19.7.1
httplib2==0.11.1
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
mccabe==0.6.1
oauth2client==4.1.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycodestyle==2.3.1
pyflakes==1.6.0
rsa==3.4.2
six==1.11.0
uritemplate==3.0.0
Werkzeug==0.14.1
